backend/requirements.txt
fastapi==0.109.1
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
aiohttp==3.13.3
pydantic==2.5.3
pydantic-settings==2.1.0
websockets==12.0
